added manifest files and supported files
added manifest files and supported files
