--- Dockerfile
@@ -1,11 +1,0 @@
-FROM python:3.9
-
-WORKDIR /app
-
-COPY . .
-
-RUN pip install --no-cache-dir -r requirements.txt
-
-EXPOSE 5000
-
-CMD ["python", "app.py"]
