--- buildspec.yml
@@ -1,41 +1,0 @@
-version: 0.2
-
-phases:
-  install:
-    runtime-versions:
-      python: 3.11
-    commands:
-      - echo Installing dependencies...
-      - pip install -r requirements.txt
-  build:
-    commands:
-      - echo "Build started on $(date)"
-      - python app.py
-artifacts:
-  files:
-    - '**/*'
-# version: 0.2
-
-# phases:
-#   install:
-#     runtime-versions:
-#       docker: latest
-#     commands:
-#       - echo Installing kubectl and awscli
-#       - curl -o kubectl https://amazon-eks.s3.ap-southeast-2.amazonaws.com/1.29.1/2024-03-18/bin/linux/amd64/kubectl
-#       - chmod +x ./kubectl
-#       - mv ./kubectl /usr/local/bin/
-#       - pip install --upgrade awscli
-#       - aws --version
-#       - echo Updating kubeconfig
-#       - aws eks update-kubeconfig --name <your-cluster-name> --region ap-southeast-2
-#   build:
-#     commands:
-#       - echo Deploying to EKS...
-#       - kubectl apply -f deployment.yaml
-#   post_build:
-#     commands:
-#       - echo Deployment complete
-# artifacts:
-#   files:
-#     - '**/*'
