+++ config/taskdef.json
@@ -4,7 +4,7 @@
   "requiresCompatibilities": ["FARGATE"],
   "cpu": "256",
   "memory": "512",
-  "executionRoleArn": "arn:aws:iam::183295435445:role/ecsTaskExecutionRole",
+  "executionRoleArn": "<EXECUTION_ROLE_ARN>",
   "containerDefinitions": [
     {
       "name": "app-container",
@@ -12,8 +12,8 @@
       "essential": true,
       "portMappings": [
         {
-          "containerPort": 80,
-          "hostPort": 80,
+          "containerPort": 8080,
+          "hostPort": 8080,
           "protocol": "tcp"
         }
       ],
