+++ docker/Dockerfile
@@ -1,0 +1,20 @@
+# Use an official Python runtime as the base image
+FROM python:3.9-slim
+
+# Set the working directory in the container
+WORKDIR /app
+
+# Copy the requirements file into the container
+COPY src/requirements.txt .
+
+# Install dependencies
+RUN pip install --no-cache-dir -r requirements.txt
+
+# Copy the rest of the application code
+COPY src/ .
+
+# Expose port 8080 (assuming your app runs on port 8080)
+EXPOSE 8080
+
+# Run the Flask app
+CMD ["python", "main.py"]
