--- requirements.txt
@@ -1,2 +1,0 @@
-Flask==2.0.1
-Werkzeug==2.0.1
