+++ src/requirements.txt
@@ -1,0 +1,2 @@
+Flask==2.0.1
+Werkzeug==2.0.1
